refactor: exception 패키지 최상단으로 이동
