+++ src/main/java/subway/adapter/in/web/exception/ExceptionAdvice.java
@@ -14,7 +14,7 @@ import org.springframework.web.bind.MissingServletRequestParameterException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import subway.application.service.exception.SubwayNoSuchResourceException;
-import subway.common.exception.SubwayIllegalArgumentException;
+import subway.exception.SubwayIllegalArgumentException;
 
 @RestControllerAdvice
 public class ExceptionAdvice {
