+++ src/main/java/subway/application/service/station/StationCommandService.java
@@ -14,7 +14,7 @@ import subway.application.port.out.line.LoadLinePort;
 import subway.application.port.out.station.LoadStationPort;
 import subway.application.port.out.station.PersistStationPort;
 import subway.application.service.exception.NoSuchStationException;
-import subway.common.exception.SubwayIllegalArgumentException;
+import subway.exception.SubwayIllegalArgumentException;
 import subway.domain.station.Station;
 
 @Service
