+++ src/main/java/subway/domain/fare/LineSurchargePolicy.java
@@ -1,7 +1,7 @@
 package subway.domain.fare;
 
 import java.util.Set;
-import subway.common.exception.SubwayIllegalArgumentException;
+import subway.exception.SubwayIllegalArgumentException;
 import subway.domain.line.Line;
 import subway.domain.route.Route;
