+++ src/main/java/subway/domain/route/RouteSection.java
@@ -1,6 +1,6 @@
 package subway.domain.route;
 
-import subway.common.exception.SubwayIllegalArgumentException;
+import subway.exception.SubwayIllegalArgumentException;
 import subway.domain.line.Line;
 import subway.domain.section.Section;
 import subway.domain.station.Station;
