+++ src/main/java/subway/domain/section/Section.java
@@ -1,6 +1,6 @@
 package subway.domain.section;
 
-import subway.common.exception.SubwayIllegalArgumentException;
+import subway.exception.SubwayIllegalArgumentException;
 import subway.domain.station.Station;
 
 public class Section {
