+++ src/main/java/subway/exception/SubwayException.java
@@ -1,4 +1,4 @@
-package subway.common.exception;
+package subway.exception;
 
 public class SubwayException extends RuntimeException {
