+++ src/test/java/subway/domain/section/SectionTest.java
@@ -8,8 +8,7 @@ import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import subway.common.exception.SubwayIllegalArgumentException;
-import subway.domain.section.Section;
+import subway.exception.SubwayIllegalArgumentException;
 import subway.fixture.StationFixture.삼성역;
 import subway.fixture.StationFixture.잠실역;
