+++ src/test/java/subway/domain/section/SectionsTest.java
@@ -11,9 +11,7 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import subway.common.exception.SubwayIllegalArgumentException;
-import subway.domain.section.Section;
-import subway.domain.section.Sections;
+import subway.exception.SubwayIllegalArgumentException;
 import subway.domain.station.Station;
 import subway.fixture.SectionFixture.이호선_역삼_삼성_3;
 import subway.fixture.StationFixture.건대역;
